Fix bugs

+++ src/main/java/com/marsmars/dtos/order/OrderItemResponse.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,5 +14,5 @@ public class OrderItemResponse {
     private Long productId;
     private String productName;
     private Integer quantity;
-    private Double priceAtOrder;
+    private BigDecimal priceAtOrder;
 }
