+++ src/main/java/com/marsmars/dtos/order/OrderResponse.java
@@ -4,6 +4,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -18,5 +19,5 @@ public class OrderResponse {
     private LocalDateTime deliveredAt;
     private Long userId;
     private List<OrderItemResponse> items;
-    private Double totalSum;
+    private BigDecimal totalSum;
 }
