+++ src/main/java/com/marsmars/models/Order.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -32,7 +33,7 @@ public class Order {
     private LocalDateTime deliveredAt;
 
     @Column(name = "total_sum")
-    private Double totalSum;
+    private BigDecimal totalSum;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
